Add the toolbox-support and toolbox-experience subpackages
These are intended to be installed into the image if the image doesn't
use the fedora-toolbox container as the base image - all the packages
are already preinstalled there.

The support subpackage should be installed to images that intent to
work flawlessly with toolbox.

The experience subpackage should be installed as well, to provide the
same experience while working in the container as one would get on the
host.

https://src.fedoraproject.org/rpms/toolbox/pull-request/1