Revert back to referring to go-md2man as go-md2man
The 'go-md2man' virtual Provides was briefly lost after the
golang-github-cpuguy83-go-md2man package was renamed to
golang-github-cpuguy83-md2man.  The virtual Provides has since been
restored [1], and go-md2man is being used as a standalone binary tool,
not as a Go package that's imported into Toolbx's source code.  Hence,
it makes sense to refer to the tool as go-md2man, and not by it's
import path.

This reverts commit 701836afcad6f0b08d8430e0914103d0871e4cdc.

There's no need to do a build just for this.

[1] golang-github-cpuguy83-md2man commit c085b15e5acd8d07
    https://src.fedoraproject.org/rpms/golang-github-cpuguy83-md2man/c/c085b15e5acd8d07